added analysis of sets over exercises of workout as well as some minor UI changes
