+++ app/src/main/java/com/example/datalift/model/Mset.kt
@@ -6,7 +6,8 @@ import com.google.firebase.firestore.IgnoreExtraProperties
 @IgnoreExtraProperties
 data class Mset(
     var rep: Long = 0,
-    var weight: Double = 0.0
+    var weight: Double = 0.0,
+    var orm: Double = 0.0
 ) {
 
     fun getFormattedSet(): String {
@@ -15,5 +16,11 @@ data class Mset(
 
     fun isValid(): Boolean {
         return rep > 0 && weight >= 0
+    }
+
+    fun setORM(){
+        if(this.rep > 0){
+            this.orm = this.weight / (1.0278 - (0.0278 * this.rep))
+        }
     }
 }
